requirements.txt
asgiref==3.5.1
astroid==2.11.6
certifi==2022.6.15
charset-normalizer==2.1.0
dill==0.3.5.1
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
googlemaps==4.6.0
greenlet==1.1.2
idna==3.3
isort==5.10.1
jedi==0.18.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
msgpack==1.0.4
parso==0.8.3
platformdirs==2.5.2
PyJWT==2.4.0
pylint==2.14.3
pynvim==0.4.3
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.0
urllib3==1.26.10
wrapt==1.14.1
